+++ dynamiq/evaluations/__init__.py
@@ -1,2 +1,3 @@
 from .base_evaluator import BaseEvaluator
+from .llm_evaluator import LLMEvaluator
 from .python_evaluator import PythonEvaluator
