+++ dynamiq/evaluations/metrics/answer_correctness.py
@@ -2,8 +2,8 @@ import json
 
 from pydantic import BaseModel, Field, PrivateAttr, field_validator, model_validator
 
-from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
 from dynamiq.evaluations import BaseEvaluator
+from dynamiq.evaluations.llm_evaluator import LLMEvaluator
 from dynamiq.nodes.llms import BaseLLM
 from dynamiq.utils.logger import logger
