+++ dynamiq/evaluations/metrics/faithfulness.py
@@ -2,8 +2,8 @@ from typing import Any
 
 from pydantic import BaseModel, PrivateAttr, field_validator, model_validator
 
-from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
 from dynamiq.evaluations import BaseEvaluator
+from dynamiq.evaluations.llm_evaluator import LLMEvaluator
 from dynamiq.nodes.llms import BaseLLM
 from dynamiq.utils.logger import logger
